\name{aqmm-internal}
\title{Internal aqmm objects}
\alias{.First.lib}
\alias{.Last.lib}
\alias{blb_part}
\alias{C_hfun_aqmm}
\alias{C_hfunD_aqmm}
\alias{getPars_gamm}
\alias{hfun_aqmm}
\alias{InitializeRE_aqmm}
\alias{loglik_aqmm}
\alias{modalRe}

\description{Internal aqmm objects.}
\details{These are not to be called by the user.}
\keyword{internal}