\name{aqmm-package}
\alias{aqmm-package}
\docType{package}
\title{
Additive Quantile Mixed Models}

\description{
Fit additive quantile mixed models
}

\details{
\tabular{ll}{
Package: \tab aqmm\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2019-03-19\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marco Geraci

Maintainer: Marco Geraci <geraci@mailbox.sc.edu>
}

\references{
Geraci M. (2014) Linear quantile mixed models: The lqmm package for Laplace quantile regression, Journal of Statistical Software, 57(13), 1--29.

Geraci, M. (2018). Additive quantile regression for clustered data with an application to children's physical activity. Journal of the Royal Statistical Society - Series C. doi:10.1111/rssc.12333 (pre-print available at \url{https://arxiv.org/abs/1803.05403}).
}

\keyword{quantile regression}
