\name{print.aqmm}
\alias{print.aqmm}

\title{
Print an \code{aqmm} Object
}
\description{
Print an object generated by \code{\link{aqmm}}.
}
\usage{
\method{print}{aqmm}(x, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{x}{
an \code{aqmm} object.
}
  \item{digits}{
a non-null value for digits specifies the minimum number of significant digits to be printed in values.
}

  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{aqmm}}
}

\keyword{print}
