\name{print.summary.aqmm}
\alias{print.summary.aqmm}

\title{
Print an \code{summary.aqmm} Object
}
\description{
Print an object generated by \code{\link{summary.aqmm}}.
}
\usage{
\method{print}{summary.aqmm}(x, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{x}{
an \code{summary.aqmm} object.
}
  \item{digits}{
a non-null value for digits specifies the minimum number of significant digits to be printed in values.
}

  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{aqmm}}
}

\keyword{print}
